+++ packages/lumx-vue/src/index.ts
@@ -26,6 +26,7 @@ export * from './components/switch';
 export * from './components/text-field';
 export * from './components/uploader';
 export * from './components/generic-block';
+export * from './components/inline-list';
 export * from './components/table';
 export * from './components/text';
 export * from './components/thumbnail';