+++ packages/site-demo/content/product/components/generic-block/vue/content-section-vertical.vue
@@ -13,26 +13,31 @@
             <Text as="p" :truncate="true" typography="body1" color="dark-L2">
                 A space to collaborate on events, communication and a workspace.
             </Text>
-            <FlexBox orientation="horizontal" gap="regular">
-                <Text as="span" typography="body1" color="dark-L2"> <Icon :icon="mdiEarth" /> Open </Text>
-                <Text as="span" typography="body1" color="dark-L2">1035 Members</Text>
-            </FlexBox>
+            <InlineList typography="body1" color="dark-L2">
+                <Text as="span"> <Icon :icon="mdiEarth" /> Open </Text>
+                <Text as="span">1035 Members</Text>
+            </InlineList>
         </template>
         <template #actions>
             <Placeholder name="3" :width="80" text-align="right" />
         </template>
     </GenericBlock>
 
-    <GenericBlock as="article" orientation="horizontal" :style="{ width: '450px' }">
+    <GenericBlock
+        as="article"
+        orientation="horizontal"
+        :style="{ width: '450px' }"
+        :actions-props="{ horizontalAlign: 'center' }"
+    >
         <template #figure>
             <Placeholder name="1" />
         </template>
         <template #default>
             <Heading as="h2" typography="subtitle1">Ines Gomez</Heading>
-            <FlexBox orientation="horizontal" gap="regular">
-                <Text as="span" typography="body1" color="dark-L2">Account Manager</Text>
-                <Text as="span" typography="body1" color="dark-L2"> Added in <Link href="#">Sales Team</Link> </Text>
-            </FlexBox>
+            <InlineList typography="body1" color="dark-L2">
+                <Text as="span">Account Manager</Text>
+                <Text as="span"> Added in <Link href="#">Sales Team</Link> </Text>
+            </InlineList>
         </template>
         <template #actions>
             <Placeholder name="3" :width="72" text-align="right" />
@@ -42,6 +47,6 @@
 
 <script setup lang="ts">
 import { mdiEarth } from '@lumx/icons';
-import { FlexBox, GenericBlock, Heading, Icon, Link, Text } from '@lumx/vue';
+import { GenericBlock, Heading, Icon, InlineList, Link, Text } from '@lumx/vue';
 import Placeholder from '@lumx/demo/components/content/Placeholder/index.vue';
 </script>